removed the single layer files, put all layers into layers.py
