+++ src/weathergen/model/layers.py
@@ -9,12 +9,116 @@
 
 
 import torch
+
+from weathergen.model.norms import AdaLayerNorm, RMSNorm
+
 from torch.utils.checkpoint import checkpoint
 
 from weathergen.model.attention import MultiSelfAttentionHead
-from weathergen.model.mlp import MLP
+# from weathergen.model.mlp import MLP
 from weathergen.model.norms import RMSNorm
 from weathergen.model.positional_encoding import positional_encoding_harmonic
+
+
+class MLP(torch.nn.Module):
+    def __init__(
+        self,
+        dim_in,
+        dim_out,
+        num_layers=2,
+        hidden_factor=2,
+        pre_layer_norm=True,
+        dropout_rate=0.0,
+        nonlin=torch.nn.GELU,
+        with_residual=False,
+        norm_type="LayerNorm",
+        dim_aux=None,
+    ):
+        """Constructor"""
+
+        super(MLP, self).__init__()
+
+        assert num_layers >= 2
+
+        self.with_residual = with_residual
+        self.with_aux = dim_aux is not None
+        dim_hidden = int(dim_in * hidden_factor)
+
+        self.layers = torch.nn.ModuleList()
+
+        norm = torch.nn.LayerNorm if norm_type == "LayerNorm" else RMSNorm
+
+        if pre_layer_norm:
+            self.layers.append(norm(dim_in) if dim_aux is None else AdaLayerNorm(dim_in, dim_aux))
+
+        self.layers.append(torch.nn.Linear(dim_in, dim_hidden))
+        self.layers.append(nonlin())
+        self.layers.append(torch.nn.Dropout(p=dropout_rate))
+
+        for _ in range(num_layers - 2):
+            self.layers.append(torch.nn.Linear(dim_hidden, dim_hidden))
+            self.layers.append(nonlin())
+            self.layers.append(torch.nn.Dropout(p=dropout_rate))
+
+        self.layers.append(torch.nn.Linear(dim_hidden, dim_out))
+        self.layers.append(nonlin())
+
+    def forward(self, *args):
+        x, x_in, aux = args[0], args[0], args[-1]
+
+        for i, layer in enumerate(self.layers):
+            x = layer(x, aux) if (i == 0 and self.with_aux) else layer(x)
+
+        if self.with_residual:
+            if x.shape[-1] == x_in.shape[-1]:
+                x = x_in + x
+            else:
+                assert x.shape[-1] % x_in.shape[-1] == 0
+                x = x + x_in.repeat([*[1 for _ in x.shape[:-1]], x.shape[-1] // x_in.shape[-1]])
+
+        return x
+
+
+class EnsPredictionHead(torch.nn.Module):
+    #########################################
+    def __init__(
+        self, dim_embed, dim_out, ens_num_layers, ens_size, norm_type="LayerNorm", hidden_factor=2
+    ):
+        """Constructor"""
+
+        super(EnsPredictionHead, self).__init__()
+
+        dim_internal = dim_embed * hidden_factor
+        # norm = torch.nn.LayerNorm if norm_type == "LayerNorm" else RMSNorm
+        enl = ens_num_layers
+
+        self.pred_heads = torch.nn.ModuleList()
+        for i in range(ens_size):
+            self.pred_heads.append(torch.nn.ModuleList())
+
+            # self.pred_heads[-1].append( norm( dim_embed))
+            self.pred_heads[-1].append(
+                torch.nn.Linear(dim_embed, dim_out if enl == 1 else dim_internal)
+            )
+
+            for i in range(ens_num_layers - 1):
+                self.pred_heads[-1].append(torch.nn.GELU())
+                self.pred_heads[-1].append(
+                    torch.nn.Linear(dim_internal, dim_out if enl - 2 == i else dim_internal)
+                )
+
+    #########################################
+    @torch.amp.custom_fwd(cast_inputs=torch.float32, device_type="cuda")
+    def forward(self, toks):
+        preds = []
+        for pred_head in self.pred_heads:
+            cpred = toks
+            for block in pred_head:
+                cpred = block(cpred)
+            preds.append(cpred)
+        preds = torch.stack(preds, 0)
+
+        return preds
 
 
 class StreamEmbedTransformer(torch.nn.Module):
@@ -169,3 +273,20 @@ class StreamEmbedTransformer(torch.nn.Module):
         out = self.dropout_final(self.ln_final(out))
 
         return out
+
+
+
+
+class StreamEmbedLinear(torch.nn.Module):
+    def __init__(self, dim_in, dim_out):
+        """Constructor"""
+
+        super(StreamEmbedLinear, self).__init__()
+
+        self.layer = torch.nn.Linear(dim_in, dim_out)
+
+    def forward(self, x):
+        # x = checkpoint( self.layer, x.flatten( -2, -1), use_reentrant=True)
+        x = self.layer(x.flatten(-2, -1))
+
+        return x
